+++ swanlab/data/run/metadata/cooperation/__init__.py
@@ -1,24 +1,24 @@
 """
 @author: cunyue
-@file: official.py
-@time: 2024/11/18 15:13
-@description: swanlab官方合作信息
+@file: __init__.py
+@time: 2024/12/3 20:17
+@description: 合作信息采集
 """
 
 import os
 
 from swanlab.api import get_http
-from swanlab.data.run.metadata.coop.qing_cloud import get_qing_cloud_info
 from swanlab.env import SwanLabEnv
 from swanlab.package import get_experiment_url
 from swanlab.package import get_package_version
+from .qing_cloud import get_qing_cloud_info
 
 
 def get_cooperation_info():
-    qing_cloud = get_qing_cloud_info()
+    qc = get_qing_cloud_info()
     coop = {"swanlab": get_swanlab_info()}
-    if qing_cloud:
-        coop.update({"qing_cloud": qing_cloud})
+    if qc:
+        coop.update({"qing_cloud": qc})
     return coop
