+++ swanlab/data/run/metadata/hardware/npu/ascend.py
@@ -5,7 +5,44 @@
 @description: 华为昇腾NPU信息采集
 """
 
+import os
+import platform
 import subprocess
+
+from swanlab.data.run.metadata.hardware.type import HardwareFuncResult
+
+
+def get_ascend_npu_info() -> HardwareFuncResult:
+    """
+    获取华为昇腾NPU信息，包括驱动版本、设备信息等
+    目前的信息统计粒度只到NPU ID级别，没有到Chip ID级别
+    """
+    # ascend芯片只支持Linux系统
+    if platform.system() != "Linux":
+        return None, []
+    # /dev目录下没有davinci*设备文件，跳过
+    # 其实理论上davinci后接数字，代表此设备id，但是官方文档也没明确写，以防万一还是不这么干了
+    if not list(filter(lambda x: x.startswith("davinci"), os.listdir("/dev"))):
+        return None, []
+    info = {"driver": None, "npu": None}
+    try:
+        # 获取NPU驱动版本
+        info["driver"] = get_version()
+        # 获取所有NPU设备ID
+        npu_map = map_npu()
+        for npu_id in npu_map:
+            for chip_id in npu_map[npu_id]:
+                chip_info = npu_map[npu_id][chip_id]
+                usage = get_chip_usage(npu_id, chip_id)
+                if info["npu"] is None:
+                    info["npu"] = {}
+                if npu_id not in info["npu"]:
+                    info["npu"][npu_id] = {}
+                info["npu"][npu_id][chip_id] = {**chip_info, "usage": usage}
+    except Exception:  # noqa
+        if all(v is None for v in info.values()):
+            return None, []
+    return info, []
 
 
 def get_version() -> str:
